requirements.txt
tensorflow>=2.13.0
opencv-python>=4.8.0
streamlit>=1.28.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
pillow>=10.0.0
kaggle>=1.5.16
tqdm>=4.66.0
lxml>=4.9.0
